pyproject.toml
[tool.poetry]
name = "cabbageapi"
version = "0.1.0"
description = ""
authors = ["Anthony Diep <antdiep14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
plaid-python = "^26.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
